pyproject.toml
[project]
name = "deep-context"
version = "0.1.0"
description = "A local AI agent for understanding entire codebases using Claude + FAISS"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "langchain>=0.2.0",
    "langchain-huggingface>=0.0.3",
    "langchain-community>=0.2.0",
    "faiss-cpu>=1.8.0",
    "tiktoken>=0.7.0",
    "rich>=13.7.0",
    "structlog>=24.1.0",
    "gitpython>=3.1.40",
    "python-dotenv>=1.0.0",
    "pathspec>=0.12.1",
    "sentence-transformers>=5.2.3",
    "langchain-anthropic>=1.3.4",
    "langchain-core>=1.2.17",
    "mcp>=1.26.0",
    "pydantic>=2.12.5",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.3.0",
]

[project.scripts]
deep-context = "src.agents.main:main"

[tool.uv]
package = true

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
